followanalytics/fa-issues#5957 Updates to the code of old templates
- Added FollowAnalyticsWrapper class to all popup templates
- Overall code improvements
