+++ popup-one-image/lib/utils.js
@@ -1,5 +1,15 @@
-export const isAndroidDevice = () => {
-  return window.navigator.platform === 'Android' ||
-         _.startsWith(window.navigator.platform, 'Linux') ||
-         window.navigator.platform === null;
+export const handleConsoleMessage = (e) => {
+  if (typeof e === 'string') {
+    console.error(`Error: ${e}`);
+  }
+  switch (e.severity) {
+    case 'warning':
+      console.warn(`Warning: ${e.message}`);
+      break;
+    case 'error':
+      console.error(`Error: ${e.message}`);
+      break;
+    default:
+      console.error(`${e}`);
+  }
 }
